+++ R/plugin-providers.R
@@ -19,10 +19,12 @@ leafletProviderDependencies <- function() {
 #'   <https://github.com/leaflet-extras/leaflet-providers>)
 #' @param layerId the layer id to assign
 #' @param group the name of the group the newly created layers should belong to
-#'   (for [clearGroup()] and [addLayersControl()] purposes).
-#'   Human-friendly group names are permitted--they need not be short,
-#'   identifier-style names.
+#'   (for [clearGroup()] and [addLayersControl()] purposes). Human-friendly
+#'   group names are permitted--they need not be short, identifier-style names.
 #' @param options tile options
+#' @param check Check that the specified `provider` matches the available
+#'   currently loaded leaflet providers? Defaults to `TRUE`, but can be toggled
+#'   to `FALSE` for advanced users.
 #' @return modified map object
 #'
 #' @examples
@@ -36,8 +38,19 @@ addProviderTiles <- function(
   provider,
   layerId = NULL,
   group = NULL,
-  options = providerTileOptions()
+  options = providerTileOptions(),
+  check = TRUE
 ) {
+  if (check) {
+    loaded_providers <- leaflet.providers::providers_loaded()
+    if (!provider %in% names(loaded_providers$providers)) {
+      stop(
+        "Unknown tile provider '",
+        provider,
+        "'; either use a known provider or pass `check = FALSE` to `addProviderTiles()`"
+      )
+    }
+  }
   map$dependencies <- c(map$dependencies, leafletProviderDependencies())
   invokeMethod(map, getMapData(map), "addProviderTiles",
     provider, layerId, group, options)